requirements.txt
click==8.0.3
joblib==1.1.0
numpy==1.22.1
pandas==1.4.0
Pillow==9.0.0
scikit-learn==1.0.2
streamlit==1.4.0
xgboost==1.4.2

